third_party/blink/web_tests/http/tests/xmlhttprequest/cache-override-expected.txt

Tests for bug 8210 - Conditional XMLHttpRequest gets should pass 304 responses unchanged.

Fetching the resource once to cache the response...
Automatically cached response body is correct; status = 200
If-Modified-Since override: status = 304
If-MODIFIED-Since override: status = 304
If-Modified-Since override (modified): status = 200
If-Unmodified-Since override: status = 200
If-Unmodified-Since override (modified): status = 412
If-Match override: status = 200
If-None-Match override: status = 304
If-None-Match override (none match): status = 200
If-Range override: status = 206
If-Range override (none match): status = 200
If-Modified-Since override (uncached response): status = 304
async: Automatically cached response body is correct; status = 200
async: If-Modified-Since override: status = 304
async: If-Modified-Since override (modified): status = 200
async: If-Unmodified-Since override: status = 200
async: If-Unmodified-Since override (modified): status = 412
async: If-Match override: status = 200
async: If-None-Match override: status = 304
async: If-None-Match override (none match): status = 200
async: If-Range override: status = 206
async: If-Range override (none match): status = 200
async: If-Modified-Since override (uncached response): status = 304
